WIP. Add a configure task to enable collectd in influxdb.conf.
Through testing I have learned (thanks Molecule!) that once collectd is
enabled, influxdb refuses to start without a file
(/usr/share/collectd/types.db) provided by the collectd package. So, I
need to figure out how to pull in the collectd role.

It looks like the way to do this is with Dependencies (thanks Gio and
Alfredo). I expect that to take a little doing, hence this WIP commit.

Note that the test is not going to pass as-is because it is so coupled
to the actual text in the config file. I have a plan for that, but I
can't actually run the test yet (the Ansible run bails due to influxdb
not starting and thus not being able to be re-started) so I'll fix it
later.

ansible-lint also hates my printf debugging tasks and I don't know why
:( but one thing at a time.