Generalize config test.
If Ansible touched the file, that's probably sufficient for this kind of
test.

In general, I think testinfra-style tests are geared to higher level
integration tests, so I'll write a better one of those next.

Consequently, I thought about deleting this test. I decided not to
because I wrote it honestly, in the process of doing TDD on this role.