Disable integration test about collectd db.
I like the idea of it but in practice the collectd database takes some
time to show up. I couldn't figure out the mechanism by which this happens,
so I'm going to delete the test and trust the influxdb.conf and systemd's
ability to complain if influx's config is actually broken.